Move application directory of user services to user/services; Refactor validate_userdata to ValidationService
